backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
google-generativeai
python-dotenv
pytz
openai
deepface
pillow
scipy
numpy
